Chapter09/cloud_functions/updateweatherpage/requirements.txt
google-cloud-datastore
google-cloud-storage
functions-framework==3.*